Condition selection with Set_exp_parameters and edit to testing valve time update

+++ Code/MATLAB and BPOD/Bpod Local/Protocols/WATER_right_ODOR_left/testing_WATER_right_ODOR_left_v2/update_valve_open_times.m
@@ -10,7 +10,7 @@ function settings_struct = update_valve_open_times(settings_struct, valve_labels
 
 arguments
     settings_struct struct
-    valve_labels int
+    valve_labels double
 end
 
 for i = 1:length(valve_labels)
